+++ projects/batfish-common-protocol/src/main/java/org/batfish/role/RoleDimensionMapping.java
@@ -1,8 +1,8 @@
 package org.batfish.role;
 
+import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkArgument;
-import static org.apache.commons.lang3.ObjectUtils.firstNonNull;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -27,7 +27,6 @@ import java.util.stream.IntStream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
-import org.batfish.common.BatfishException;
 
 /**
  * Objects of this class represent a way to map node names to role names for a particular role
@@ -65,7 +64,8 @@ public class RoleDimensionMapping implements Comparable<RoleDimensionMapping> {
     try {
       _pattern = Pattern.compile(regex);
     } catch (PatternSyntaxException e) {
-      throw new BatfishException("Supplied regex is not a valid Java regex: \"" + regex + "\"", e);
+      throw new IllegalArgumentException(
+          "Supplied regex is not a valid Java regex: \"" + regex + "\"", e);
     }
     _groups =
         firstNonNull(
@@ -165,7 +165,7 @@ public class RoleDimensionMapping implements Comparable<RoleDimensionMapping> {
       try {
         p = Pattern.compile(roleRegex);
       } catch (PatternSyntaxException e) {
-        throw new BatfishException(
+        throw new IllegalArgumentException(
             "Supplied regex is not a valid Java regex: \"" + roleRegex + "\"", e);
       }
       if (p.matcher(roleName).matches()) {
