+++ projects/common/src/main/java/BUILD
@@ -2,11 +2,8 @@ package(default_visibility = ["//visibility:public"])
 
 java_library(
     name = "common",
-    srcs = glob(["**/*.java"]),
-    deps = [
-        "@maven//:com_fasterxml_jackson_core_jackson_annotations",
-        "@maven//:com_github_ben_manes_caffeine_caffeine",
-        "@maven//:com_google_code_findbugs_jsr305",
-        "@maven//:com_google_guava_guava",
+    exports = [
+        "//projects/common/src/main/java/org/batfish/datamodel:action",
+        "//projects/common/src/main/java/org/batfish/datamodel:ipv4",
     ],
 )
