+++ projects/common/src/main/java/org/batfish/common/util/BUILD
@@ -1,0 +1,10 @@
+package(default_visibility = ["//visibility:public"])
+
+java_library(
+    name = "interface_name_comparator",
+    srcs = [":InterfaceNameComparator.java"],
+    deps = [
+        "@maven//:com_google_code_findbugs_jsr305",
+        "@maven//:com_google_guava_guava",
+    ],
+)
